ml_services/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1

pandas==2.2.2
numpy==2.0.2
scikit-learn==1.5.2
requests==2.32.3

# Gemini AI integration
google-generativeai==0.3.2
python-dotenv==1.0.0

# Geo stack (needed for soilgrids_districts.py)
geopandas==1.0.1
shapely==2.0.6
pyproj==3.6.1
fiona==1.9.6

# Disease Prediction Module
tensorflow==2.13.0
Pillow>=9.0.0
